Initial Codebase
Small Batch file to get the official Logomotion code from 2011, and the
last fully featured Elot codebase that watched for Kinect telemetry.
This is essentially the base of the project.
(The code here was written by the Force Team 1073 in 2011 - I am merely
committing the code, as I am one of many authors of this frc project)